Relocate VSM/HEKI service code into mshv::heki_service
Split the HEKI service logic out of the monolithic mshv/vsm.rs into a new
mshv::heki_service submodule (vsm, state, mem_integrity). Pure relocation: no
logic, signature, or behavior changes. The platform vsm.rs retains only VTL
bring-up, control-register locking, the vsm_dispatch entry point, and the
protected-frame registry (with FrameReservation). Cross-module visibility is
widened minimally where the relocated service handlers reach into the registry;
a follow-up introduces the VsmBackend abstraction that removes that coupling.
